MoQ support (#468)
* LOC WIP

* Move to SwiftLintPlugins version of plugin

* wip

* LOC e2e

* Cleanup

* Dates are already absolute

* WIP

* WIP

* Remove QMedia ObjectiveC

* Remove QMedia

* WIP

* Audio working

* Fixup tests

* WIP

* Last day

* Fixup retroactive and LOC

* Make libquicr SHARED

* Restore platforms

* Docs

* Add license header autogen

* More license updates

* Deinit watching

* Add TestFlight link

* swiftlint fixups

* LOC E2E

+++ Decimus/QHelpers.swift
@@ -3,19 +3,19 @@
 
 /// Provides handling of names.
 protocol NameGate {
-    func handle(groupId: UInt32, objectId: UInt16, lastGroup: UInt32?, lastObject: UInt16?) -> Bool
+    func handle(groupId: UInt64, objectId: UInt64, lastGroup: UInt64?, lastObject: UInt64?) -> Bool
 }
 
 /// A NameGate that allows anything.
 class AllowAllNameGate: NameGate {
-    func handle(groupId: UInt32, objectId: UInt16, lastGroup: UInt32?, lastObject: UInt16?) -> Bool {
+    func handle(groupId: UInt64, objectId: UInt64, lastGroup: UInt64?, lastObject: UInt64?) -> Bool {
         true
     }
 }
 
 /// A NameGate that only allows sequential object inside a group, and incrementing groups when their object == 0.
 class SequentialObjectBlockingNameGate: NameGate {
-    func handle(groupId: UInt32, objectId: UInt16, lastGroup: UInt32?, lastObject: UInt16?) -> Bool {
+    func handle(groupId: UInt64, objectId: UInt64, lastGroup: UInt64?, lastObject: UInt64?) -> Bool {
         // 0th object of a newer group can always be written.
         if (lastGroup == nil || groupId > lastGroup!) && objectId == 0 {
             return true
